add accessible check to ajax_function

+++ library/Admin/library/Admin/Component/VenueAliasList.php
@@ -18,6 +18,10 @@ class Admin_Component_VenueAliasList extends Admin_Component_Abstract {
 	}
 
 	public static function ajax_deleteAlias(CM_Params $params, CM_ComponentFrontendHandler $handler, CM_Response_View_Ajax $response) {
+		if (!$response->getViewer(true)->getRoles()->contains(Denkmal_Role::ADMIN)) {
+			throw new CM_Exception_AuthRequired();
+		}
+
 		/** @var Denkmal_Params $params */
 		$venueAlias = $params->getVenueAlias('id');
 		$venueAlias->delete();
